move merkle tree implementation
